# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flowers Shop Documentation API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/flowers/{flower}": {
            "get": {
                "tags": [
                    "Flower"
                ],
                "summary": "Цветок по ID",
                "operationId": "c9103ec6b19fb2135ce29e778cdeafde",
                "parameters": [
                    {
                        "name": "flower",
                        "in": "path",
                        "description": "ID цветка",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Title"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Description"
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 199
                                                },
                                                "category": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Title category"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "tags": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Title tags"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "images": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Title image"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Flower"
                ],
                "summary": "Удалить",
                "operationId": "7f6038c2e1cba65be59651c2a4846ab2",
                "parameters": [
                    {
                        "name": "flower",
                        "in": "path",
                        "description": "ID цветка",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Flower"
                ],
                "summary": "Обновить",
                "operationId": "316670d4fe2a4f7fadad707e8376b447",
                "parameters": [
                    {
                        "name": "flower",
                        "in": "path",
                        "description": "ID цветка",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Title is update"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Description"
                                            },
                                            "count": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 8
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "description": "The tags ID",
                                                    "type": "number"
                                                }
                                            },
                                            "is_published": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flowers/": {
            "post": {
                "tags": [
                    "Flower"
                ],
                "summary": "Создать",
                "operationId": "5ed746d086a165f36792624abf5783a6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Title is create"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Description"
                                            },
                                            "count": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": 199
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "description": "The tags ID",
                                                    "type": "number"
                                                }
                                            },
                                            "is_published": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}